fix: renames file as workarround for name clashes with QMake
